+++ src/pixelator/pna/pixeldataset/io/pxl_file.py
@@ -11,6 +11,8 @@ from pathlib import Path
 
 import duckdb
 
+from pixelator.common.duckdb_utils import connect_duckdb
+
 PXL_FILE_MANDATOR_TABLES = [
     "__adata__X",
     "__adata__var",
@@ -48,7 +50,7 @@ class PxlFile:
 
     def is_pxl_file(self) -> bool:
         """Check if the file is a PXL file."""
-        with duckdb.connect(self.path, read_only=True) as con:
+        with connect_duckdb(self.path, read_only=True) as con:
             tables = con.sql("SHOW ALL TABLES").to_df()
             return len(
                 set(PXL_FILE_MANDATOR_TABLES).intersection(
@@ -59,7 +61,7 @@ class PxlFile:
     def metadata(self) -> dict:
         """Read the metadata from the PXL file."""
         try:
-            with duckdb.connect(self.path, read_only=True) as con:
+            with connect_duckdb(self.path, read_only=True) as con:
                 metadata = con.sql("SELECT * FROM metadata").fetchone()
                 return json.loads(metadata[0]) if metadata else {}
         except duckdb.CatalogException: